requirements.txt
black==22.12.0
py-cord==2.3.2
gspread==5.7.2
mysql-connector-python==8.0.32
python-dotenv==0.21.1
pytz==2022.7.1